chore: design parity
